requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy[ext]
asyncpg
alembic
python-multipart
python-dotenv
pytest
httpx
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
aiosqlite
pytest-asyncio
